+++ CHANGELOG.md
@@ -5,4 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [unreleased]
+
+### Fixed
+- Fixed the issue where a ticket could be solved without a solution
+
 ## [1.0.0-rc1]